Minor updates
Added neutral and the transition to happy video's from Davey. Though they are not the final video's and have to have a lot of smoothing.

Changed the "message" function to the "look" function. The "message" function can now be used to do more than one thing, but it is still in progress so there is just a console log.